pyproject.toml
[project]
name = "my-django-rest-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.6",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "drf-yasg>=1.21.11",
    "inflection>=0.5.1",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.11.0",
    "pyyaml>=6.0.3",
]
